// src/app/modules/logistica/yms/setores/models/setores.ts
export interface ILogisticaYmsSetores {
  UUID_LOGI_YMS_SETO: number;
  UUID_LOGI_YMS_SETO_TIPO: number;
  NM_SETO_TIPO: string;
  UUID_LOGI_FILI: number;
  NM_SETO: string;
  DS_ICON: string;
  NM_FILI: string;
  DS_OBSE: string;
  ID_USUA: number;
  NR_MATR: number;
  NM_USUA: string;
  IP_USUA: string;
  IN_STAT: number;
  DT_ATUA: string;
  DT_INCL: string;
  IN_PAGI: number;
}